Make double quotes the default string type
There is an article found here:
http://viget.com/extend/just-use-double-quoted-ruby-strings
That raised some interesting points. The main reason for using single
quotes as a default, and only using double quotes when interpolation is
needed stems from the argument that double quotes require a higher
performance hit to process. This article seems to squash that theory,
along with a few others.